+++ packages/jsonapi/tests/data/ApiTypes/EmptyType.php
@@ -9,15 +9,9 @@ use EDT\JsonApi\ApiDocumentation\AttributeTypeResolver;
 use EDT\JsonApi\RequestHandling\MessageFormatter;
 use EDT\JsonApi\ResourceTypes\AbstractResourceType;
 use EDT\JsonApi\ResourceTypes\PropertyBuilderFactory;
-use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Querying\Contracts\PropertyAccessorInterface;
-use EDT\Querying\PropertyAccessors\ReflectionPropertyAccessor;
 use EDT\Querying\PropertyPaths\PropertyPath;
-use EDT\Querying\Utilities\ConditionEvaluator;
-use EDT\Querying\Utilities\Sorter;
-use EDT\Querying\Utilities\TableJoiner;
 use EDT\Wrapping\Contracts\EntityFetcherInterface;
-use EDT\Wrapping\Contracts\Types\ReindexableTypeInterface;
 use EDT\Wrapping\Utilities\SchemaPathProcessor;
 use Psr\Log\LoggerInterface;
 use Tests\data\EmptyEntity;
@@ -96,6 +90,11 @@ class EmptyType extends AbstractResourceType
     }
 
     protected function getEntityFetcher(): EntityFetcherInterface
+    {
+        throw new \RuntimeException();
+    }
+
+    protected function getIdentifierPropertyPath(): array
     {
         throw new \RuntimeException();
     }